fake-store-backend/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-CORS
firebase-admin
psycopg2-binary
python-dotenv
